Add files via upload
The technology firm has 3 potential hackers that perpetrated the attack. 
They are certain of the first two hackers but they aren't very sure if the third hacker was involved or not. 
They have requested your help! 
Can you help figure out whether or not the third suspect had anything to do with the attacks, or was it just two hackers?